Optimized conversion terms and renamed file
